editor/libeditor/tests/test_nsITableEditor_deleteTableColumn.html

<!DOCTYPE>
<html>
<head>
  <title>Test for nsITableEditor.deleteTableColumn()</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" href="/tests/SimpleTest/test.css">
</head>
<body>
<div id="display">
</div>
<div id="content" contenteditable>out of table<table><tr><td>default content</td></tr></table></div>
<pre id="test">
</pre>

<script class="testbody" type="application/javascript">
"use strict";

SimpleTest.waitForExplicitFinish();
SimpleTest.waitForFocus(async () => {
  await SpecialPowers.pushPrefEnv({
    set: [["dom.input_events.beforeinput.enabled", true]],
  });

  let editor = document.getElementById("content");
  let selection = document.getSelection();

  function checkInputEvent(aEvent, aDescription) {
    ok(aEvent instanceof InputEvent,
       `"${aEvent.type}" event should be dispatched with InputEvent interface ${aDescription}`);
    is(aEvent.cancelable, aEvent.type === "beforeinput",
       `"${aEvent.type}" event should ${aEvent.type === "beforeinput" ? "be" : "be never"} cancelable ${aDescription}`);
    is(aEvent.bubbles, true,
       `"${aEvent.type}" event should always bubble ${aDescription}`);
    is(aEvent.inputType, "deleteContent",
       `inputType of "${aEvent.type}" event should be "deleteContent" ${aDescription}`);
    is(aEvent.data, null,
       `data of "${aEvent.type}" event should be null ${aDescription}`);
    is(aEvent.dataTransfer, null,
       `dataTransfer of "${aEvent.type}" event should be null ${aDescription}`);
  }

  let beforeInputEvents = [];
  let inputEvents = [];
  function onBeforeInput(aEvent) {
    beforeInputEvents.push(aEvent);
  }
  function onInput(aEvent) {
    inputEvents.push(aEvent);
  }
  editor.addEventListener("beforeinput", onBeforeInput);
  editor.addEventListener("input", onInput);

  beforeInputEvents = [];
  inputEvents = [];
  selection.collapse(editor.firstChild, 0);
  getTableEditor().deleteTableColumn(1);
  is(editor.innerHTML, "out of table<table><tbody><tr><td>default content</td></tr></tbody></table>",
     "nsITableEditor.deleteTableColumn(1) should do nothing if selection is not in <table>");
  // If there were specific inputType value for this API, we should dispatch cancelable "beforeinput", though.
  todo_is(beforeInputEvents.length, 1,
     '"beforeinput" event should be fired when a call of nsITableEditor.deleteTableColumn(1) will do nothing');
  if (beforeInputEvents.length > 0) {
    checkInputEvent(beforeInputEvents[0], "when selection is collapsed outside table element (nsITableEditor.deleteTableColumn(1))");
  }
  is(inputEvents.length, 0,
     'No "input" event should be fired when a call of nsITableEditor.deleteTableColumn(1) does nothing');

  selection.removeAllRanges();
  try {
    beforeInputEvents = [];
    inputEvents = [];
    getTableEditor().deleteTableColumn(1);
    ok(false, "getTableEditor().deleteTableColumn(1) without selection ranges should throw exception");
  } catch (e) {
    ok(true, "getTableEditor().deleteTableColumn(1) without selection ranges should throw exception");
    is(beforeInputEvents.length, 0,
       'No "beforeinput" event should be fired when nsITableEditor.deleteTableColumn(1) causes exception due to no selection range');
    is(inputEvents.length, 0,
       'No "input" event should be fired when nsITableEditor.deleteTableColumn(1) causes exception due to no selection range');
  }

  // If a cell is selected and the argument is less than number of rows,
  // specified number of rows should be removed starting from the row
  // containing the selected cell.  But if the argument is same or
  // larger than actual number of rows, the <table> should be removed.
  selection.removeAllRanges();
  editor.innerHTML =
    '<table><tr><td id="select">cell1-1</td><td>cell1-2</td></tr><tr><td>cell2-1</td><td>cell2-2</td></tr></table>';
  beforeInputEvents = [];
  inputEvents = [];
  let range = document.createRange();
  range.selectNode(document.getElementById("select"));
  selection.addRange(range);
  getTableEditor().deleteTableColumn(1);
  is(editor.innerHTML, "<table><tbody><tr><td>cell1-2</td></tr><tr><td>cell2-2</td></tr></tbody></table>",
     "nsITableEditor.deleteTableColumn(1) should delete the first column when a cell in the first column is selected");
  todo_is(beforeInputEvents.length, 1,
     'Only one "beforeinput" event should be fired when a cell in the first column is selected');
  if (beforeInputEvents.length > 0) {
    checkInputEvent(beforeInputEvents[0], "when a cell in the first column is selected");
  }
  is(inputEvents.length, 1,
     'Only one "input" event should be fired when a cell in the first column is selected');
  checkInputEvent(inputEvents[0], "when a cell in the first column is selected");

  selection.removeAllRanges();
  editor.innerHTML =
    '<table><tr><td>cell1-1</td><td>cell1-2</td></tr><tr><td>cell2-1</td><td id="select">cell2-2</td></tr></table>';
  beforeInputEvents = [];
  inputEvents = [];
  range = document.createRange();
  range.selectNode(document.getElementById("select"));
  selection.addRange(range);
  getTableEditor().deleteTableColumn(1);
  is(editor.innerHTML, "<table><tbody><tr><td>cell1-1</td></tr><tr><td>cell2-1</td></tr></tbody></table>",
     "nsITableEditor.deleteTableColumn(1) should delete the second column when a cell in the second column is selected");
  todo_is(beforeInputEvents.length, 1,
     'Only one "beforeinput" event should be fired when a cell in the second column is selected');
  if (beforeInputEvents.length > 0) {
    checkInputEvent(beforeInputEvents[0], "when a cell in the second column is selected");
  }
  is(inputEvents.length, 1,
     'Only one "input" event should be fired when a cell in the second column is selected');
  checkInputEvent(inputEvents[0], "when a cell in the second column is selected");

  selection.removeAllRanges();
  editor.innerHTML =
    '<table><tr><td id="select">cell1-1</td><td>cell1-2</td></tr><tr><td>cell2-1</td><td>cell2-2</td></tr></table>';
  beforeInputEvents = [];
  inputEvents = [];
  range = document.createRange();
  range.selectNode(document.getElementById("select"));
  selection.addRange(range);
  getTableEditor().deleteTableColumn(2);
  is(editor.innerHTML, "",
     "nsITableEditor.deleteTableColumn(2) should delete the <table> since there is only 2 columns");
  todo_is(beforeInputEvents.length, 1,
     'Only one "beforeinput" event should be fired when a cell in first column is selected and argument is same as number of rows');
  if (beforeInputEvents.length > 0) {
    checkInputEvent(beforeInputEvents[0], "when a cell in first column is selected and argument is same as number of rows");
  }
  is(inputEvents.length, 1,
     'Only one "input" event should be fired when a cell in first column is selected and argument is same as number of rows');
  checkInputEvent(inputEvents[0], "when a cell in first column is selected and argument is same as number of rows");

  selection.removeAllRanges();
  editor.innerHTML =
    '<table><tr><td id="select">cell1-1</td><td>cell1-2</td></tr><tr><td>cell2-1</td><td>cell2-2</td></tr></table>';
  beforeInputEvents = [];
  inputEvents = [];
  range = document.createRange();
  range.selectNode(document.getElementById("select"));
  selection.addRange(range);
  getTableEditor().deleteTableColumn(3);
  is(editor.innerHTML, "",
     "nsITableEditor.deleteTableColumn(3) should delete the <table> when argument is larger than actual number of columns");
  todo_is(beforeInputEvents.length, 1,
     'Only one "beforeinput" event should be fired when argument is larger than actual number of columns');
  if (beforeInputEvents.length > 0) {
    checkInputEvent(beforeInputEvents[0], "when argument is larger than actual number of columns");
  }
  is(inputEvents.length, 1,
     'Only one "input" event should be fired when argument is larger than actual number of columns');
  checkInputEvent(inputEvents[0], "when argument is larger than actual number of columns");

  selection.removeAllRanges();
  editor.innerHTML =
    '<table><tr><td>cell1-1</td><td id="select">cell1-2</td><td>cell1-3</td></tr><tr><td>cell2-1</td><td>cell2-2</td><td>cell2-3</td></tr></table>';
  beforeInputEvents = [];
  inputEvents = [];
  range = document.createRange();
  range.selectNode(document.getElementById("select"));
  selection.addRange(range);
  getTableEditor().deleteTableColumn(2);
  is(editor.innerHTML, "<table><tbody><tr><td>cell1-1</td></tr><tr><td>cell2-1</td></tr></tbody></table>",
     "nsITableEditor.deleteTableColumn(2) should delete the second column containing selected cell and next column");
  todo_is(beforeInputEvents.length, 1,
     'Only one "beforeinput" event should be fired when a cell in second column and argument is same as the remaining columns');
  if (beforeInputEvents.length > 0) {
    checkInputEvent(beforeInputEvents[0], "when a cell in second column and argument is same as the remaining columns");
  }
  is(inputEvents.length, 1,
     'Only one "input" event should be fired when a cell in second column and argument is same as the remaining columns');
  checkInputEvent(inputEvents[0], "when a cell in second column and argument is same as the remaining columns");

  selection.removeAllRanges();
  editor.innerHTML =
    '<table><tr><td>cell1-1</td><td>cell1-2</td></tr><tr><td id="select">cell2-1</td><td>cell2-2</td></tr><tr><td>cell3-1</td><td>cell3-2</td></tr></table>';
  beforeInputEvents = [];
  inputEvents = [];
  range = document.createRange();
  range.selectNode(document.getElementById("select"));
  selection.addRange(range);
  getTableEditor().deleteTableColumn(3);
  is(editor.innerHTML, "",
     "nsITableEditor.deleteTableColumn(3) should delete the <table> since the argument equals actual number of columns");
  todo_is(beforeInputEvents.length, 1,
     'Only one "beforeinput" event should be fired when a cell in first column and argument is larger than the remaining columns');
  if (beforeInputEvents.length > 0) {
    checkInputEvent(beforeInputEvents[0], "when a cell in first column and argument is larger than the remaining columns");
  }
  is(inputEvents.length, 1,
     'Only one "input" event should be fired when a cell in first column and argument is larger than the remaining columns');
  checkInputEvent(inputEvents[0], "when a cell in first column and argument is larger than the remaining columns");

  // Similar to selected a cell, when selection is in a cell, the cell should
  // treated as selected.
  selection.removeAllRanges();
  editor.innerHTML =
    '<table><tr><td id="select">cell1-1</td><td>cell1-2</td></tr><tr><td>cell2-1</td><td>cell2-2</td></tr></table>';
  beforeInputEvents = [];
  inputEvents = [];
  editor.scrollTop; // Needs to flush pending reflow since we need layout information in this case.
  range = document.createRange();
  range.selectNode(document.getElementById("select").firstChild);
  selection.addRange(range);
  getTableEditor().deleteTableColumn(1);
  is(editor.innerHTML, "<table><tbody><tr><td>cell1-2</td></tr><tr><td>cell2-2</td></tr></tbody></table>",
     "nsITableEditor.deleteTableColumn(1) should delete the first column when a cell in the first column contains selection range");
  todo_is(beforeInputEvents.length, 1,
     'Only one "beforeinput" event should be fired when a cell in the first column contains selection range');
  if (beforeInputEvents.length > 0) {
    checkInputEvent(beforeInputEvents[0], "when a cell in the first column contains selection range");
  }
  is(inputEvents.length, 1,
     'Only one "input" event should be fired when a cell in the first column contains selection range');
  checkInputEvent(inputEvents[0], "when a cell in the first column contains selection range");

  selection.removeAllRanges();
  editor.innerHTML =
    '<table><tr><td>cell1-1</td><td>cell1-2</td></tr><tr><td>cell2-1</td><td id="select">cell2-2</td></tr></table>';
  beforeInputEvents = [];
  inputEvents = [];
  editor.scrollTop; // Needs to flush pending reflow since we need layout information in this case.
  range = document.createRange();
  range.selectNode(document.getElementById("select").firstChild);
  selection.addRange(range);
  getTableEditor().deleteTableColumn(1);
  is(editor.innerHTML, "<table><tbody><tr><td>cell1-1</td></tr><tr><td>cell2-1</td></tr></tbody></table>",
     "nsITableEditor.deleteTableColumn(1) should delete the second column when a cell in the second column contains selection range");
  todo_is(beforeInputEvents.length, 1,
     'Only one "beforeinput" event should be fired when a cell in the second column contains selection range');
  if (beforeInputEvents.length > 0) {
    checkInputEvent(beforeInputEvents[0], "when a cell in the second column contains selection range");
  }
  is(inputEvents.length, 1,
     'Only one "input" event should be fired when a cell in the second column contains selection range');
  checkInputEvent(inputEvents[0], "when a cell in the second column contains selection range");

  selection.removeAllRanges();
  editor.innerHTML =
    '<table><tr><td id="select">cell1-1</td><td>cell1-2</td></tr><tr><td>cell2-1</td><td>cell2-2</td></tr></table>';
  beforeInputEvents = [];
  inputEvents = [];
  editor.scrollTop; // Needs to flush pending reflow since we need layout information in this case.
  range = document.createRange();
  range.selectNode(document.getElementById("select").firstChild);
  selection.addRange(range);
  getTableEditor().deleteTableColumn(2);
  is(editor.innerHTML, "",
     "nsITableEditor.deleteTableColumn(2) should delete the <table> since there is only 2 columns");
  todo_is(beforeInputEvents.length, 1,
     'Only one "beforeinput" event should be fired when all text in a cell in first column is selected and argument includes next row');
  if (beforeInputEvents.length > 0) {
    checkInputEvent(beforeInputEvents[0], "when all text in a cell in first column is selected and argument includes next row");
  }
  is(inputEvents.length, 1,
     'Only one "input" event should be fired when all text in a cell in first column is selected and argument includes next row');
  checkInputEvent(inputEvents[0], "when all text in a cell in first column is selected and argument includes next row");

  selection.removeAllRanges();
  editor.innerHTML =
    '<table><tr><td id="select">cell1-1</td><td>cell1-2</td></tr><tr><td>cell2-1</td><td>cell2-2</td></tr></table>';
  beforeInputEvents = [];
  inputEvents = [];
  editor.scrollTop; // Needs to flush pending reflow since we need layout information in this case.
  range = document.createRange();
  range.selectNode(document.getElementById("select").firstChild);
  selection.addRange(range);
  getTableEditor().deleteTableColumn(3);
  is(editor.innerHTML, "",
     "nsITableEditor.deleteTableColumn(3) should delete the <table> when argument is larger than actual number of columns");
  todo_is(beforeInputEvents.length, 1,
     'Only one "beforeinput" event should be fired when all text in a cell in first column is selected and argument is same as number of all rows');
  if (beforeInputEvents.length > 0) {
    checkInputEvent(beforeInputEvents[0], "when all text in a cell in first column is selected and argument is same as number of all rows");
  }
  is(inputEvents.length, 1,
     'Only one "input" event should be fired when all text in a cell in first column is selected and argument is same as number of all rows');
  checkInputEvent(inputEvents[0], "when all text in a cell in first column is selected and argument is same as number of all rows");

  selection.removeAllRanges();
  editor.innerHTML =
    '<table><tr><td>cell1-1</td><td id="select">cell1-2</td><td>cell1-3</td></tr><tr><td>cell2-1</td><td>cell2-2</td><td>cell2-3</td></tr></table>';
  beforeInputEvents = [];
  inputEvents = [];
  editor.scrollTop; // Needs to flush pending reflow since we need layout information in this case.
  range = document.createRange();
  range.selectNode(document.getElementById("select").firstChild);
  selection.addRange(range);
  getTableEditor().deleteTableColumn(2);
  is(editor.innerHTML, "<table><tbody><tr><td>cell1-1</td></tr><tr><td>cell2-1</td></tr></tbody></table>",
     "nsITableEditor.deleteTableColumn(2) should delete the second column containing a cell containing selection range and next column");
  todo_is(beforeInputEvents.length, 1,
     'Only one "beforeinput" event should be fired when all text in a cell is selected and argument is same than renaming number of columns');
  if (beforeInputEvents.length > 0) {
    checkInputEvent(beforeInputEvents[0], "when all text in a cell is selected and argument is same than renaming number of columns");
  }
  is(inputEvents.length, 1,
     'Only one "input" event should be fired when all text in a cell is selected and argument is same than renaming number of columns');
  checkInputEvent(inputEvents[0], "when all text in a cell is selected and argument is same than renaming number of columns");

  selection.removeAllRanges();
  editor.innerHTML =
    '<table><tr><td>cell1-1</td><td>cell1-2</td></tr><tr><td id="select">cell2-1</td><td>cell2-2</td></tr><tr><td>cell3-1</td><td>cell3-2</td></tr></table>';
  beforeInputEvents = [];
  inputEvents = [];
  editor.scrollTop; // Needs to flush pending reflow since we need layout information in this case.
  range = document.createRange();
  range.selectNode(document.getElementById("select").firstChild);
  selection.addRange(range);
  getTableEditor().deleteTableColumn(3);
  is(editor.innerHTML, "",
     "nsITableEditor.deleteTableColumn(3) should delete the <table> since the argument equals actual number of columns");
  todo_is(beforeInputEvents.length, 1,
     'Only one "beforeinput" event should be fired when all text in a cell in the first column and argument is larger than renaming number of columns');
  if (beforeInputEvents.length > 0) {
    checkInputEvent(beforeInputEvents[0], "when all text in a cell in the first column and argument is larger than renaming number of columns");
  }
  is(inputEvents.length, 1,
     'Only one "input" event should be fired when all text in a cell in the first column and argument is larger than renaming number of columns');
  checkInputEvent(inputEvents[0], "when all text in a cell in the first column and argument is larger than renaming number of columns");

  // The argument should be ignored when 2 or more cells are selected.
  // XXX Different from deleteTableRow(), this removes the <table> completely.
  selection.removeAllRanges();
  editor.innerHTML =
    '<table><tr><td id="select1">cell1-1</td><td>cell1-2</td></tr><tr><td>cell2-1</td><td id="select2">cell2-2</td></tr></table>';
  beforeInputEvents = [];
  inputEvents = [];
  range = document.createRange();
  range.selectNode(document.getElementById("select1"));
  selection.addRange(range);
  range = document.createRange();
  range.selectNode(document.getElementById("select2"));
  selection.addRange(range);
  getTableEditor().deleteTableColumn(1);
  is(editor.innerHTML, "",
     "nsITableEditor.deleteTableColumn(1) should delete the <table> when both columns have selected cell");
  todo_is(beforeInputEvents.length, 1,
     'Only one "beforeinput" event should be fired when both columns have selected cell');
  if (beforeInputEvents.length > 0) {
    checkInputEvent(beforeInputEvents[0], "when both columns have selected cell");
  }
  is(inputEvents.length, 1,
     'Only one "input" event should be fired when both columns have selected cell');
  checkInputEvent(inputEvents[0], "when both columns have selected cell");

  selection.removeAllRanges();
  editor.innerHTML =
    '<table><tr><td id="select1">cell1-1</td><td>cell1-2</td></tr><tr><td id="select2">cell2-1</td><td>cell2-2</td></tr></table>';
  beforeInputEvents = [];
  inputEvents = [];
  range = document.createRange();
  range.selectNode(document.getElementById("select1"));
  selection.addRange(range);
  range = document.createRange();
  range.selectNode(document.getElementById("select2"));
  selection.addRange(range);
  getTableEditor().deleteTableColumn(2);
  is(editor.innerHTML, "",
     "nsITableEditor.deleteTableColumn(2) should delete the <table> since 2 is number of columns of the <table>");
  todo_is(beforeInputEvents.length, 1,
     'Only one "beforeinput" event should be fired when cells in every column are selected #2');
  if (beforeInputEvents.length > 0) {
    checkInputEvent(beforeInputEvents[0], "when cells in every column are selected #2");
  }
  is(inputEvents.length, 1,
     'Only one "input" event should be fired when cells in every column are selected #2');
  checkInputEvent(inputEvents[0], "when cells in every column are selected #2");

  selection.removeAllRanges();
  editor.innerHTML =
    '<table><tr><td id="select1">cell1-1</td><td>cell1-2</td></tr><tr><td id="select2">cell2-1</td><td>cell2-2</td></tr></table>';
  beforeInputEvents = [];
  inputEvents = [];
  range = document.createRange();
  range.selectNode(document.getElementById("select1"));
  selection.addRange(range);
  range = document.createRange();
  range.selectNode(document.getElementById("select2"));
  selection.addRange(range);
  getTableEditor().deleteTableColumn(2);
  is(editor.innerHTML, "",
     "nsITableEditor.deleteTableColumn(2) should delete the <table> since 2 is number of columns of the <table>");
  todo_is(beforeInputEvents.length, 1,
     'Only one "beforeinput" event should be fired when 2 cells in same column are selected');
  if (beforeInputEvents.length > 0) {
    checkInputEvent(beforeInputEvents[0], "when 2 cells in same column are selected");
  }
  is(inputEvents.length, 1,
     'Only one "input" event should be fired when 2 cells in same column are selected');
  checkInputEvent(inputEvents[0], "when 2 cells in same column are selected");

  selection.removeAllRanges();
  editor.innerHTML =
    '<table><tr><td id="select1">cell1-1</td><td id="select2">cell1-2</td><td>cell1-3</td></tr><tr><td>cell2-1</td><td>cell2-2</td><td>cell2-3</td></tr></table>';
  beforeInputEvents = [];
  inputEvents = [];
  range = document.createRange();
  range.selectNode(document.getElementById("select1"));
  selection.addRange(range);
  range = document.createRange();
  range.selectNode(document.getElementById("select2"));
  selection.addRange(range);
  getTableEditor().deleteTableColumn(1);
  is(editor.innerHTML, "<table><tbody><tr><td>cell1-3</td></tr><tr><td>cell2-3</td></tr></tbody></table>",
     "nsITableEditor.deleteTableColumn(1) should delete first 2 columns because cells in the both columns are selected");
  todo_is(beforeInputEvents.length, 1,
     'Only one "beforeinput" event should be fired when 2 cell elements in different columns are selected #1');
  if (beforeInputEvents.length > 0) {
    checkInputEvent(beforeInputEvents[0], "when 2 cell elements in different columns are selected #1");
  }
  is(inputEvents.length, 1,
     'Only one "input" event should be fired when 2 cell elements in different columns are selected #1');
  checkInputEvent(inputEvents[0], "when 2 cell elements in different columns are selected #1");

  selection.removeAllRanges();
  editor.innerHTML =
    '<table><tr><td id="select1">cell1-1</td><td>cell1-2</td><td id="select2">cell1-3</td></tr><tr><td>cell2-1</td><td>cell2-2</td><td>cell2-3</td></tr></table>';
  beforeInputEvents = [];
  inputEvents = [];
  range = document.createRange();
  range.selectNode(document.getElementById("select1"));
  selection.addRange(range);
  range = document.createRange();
  range.selectNode(document.getElementById("select2"));
  selection.addRange(range);
  getTableEditor().deleteTableColumn(1);
  is(editor.innerHTML, "<table><tbody><tr><td>cell1-2</td></tr><tr><td>cell2-2</td></tr></tbody></table>",
     "nsITableEditor.deleteTableColumn(1) should delete the first and the last columns because cells in the both columns are selected");
  todo_is(beforeInputEvents.length, 1,
     'Only one "beforeinput" event should be fired when 2 cell elements in different columns are selected #2');
  if (beforeInputEvents.length > 0) {
    checkInputEvent(beforeInputEvents[0], "when 2 cell elements in different columns are selected #2");
  }
  is(inputEvents.length, 1,
     'Only one "input" event should be fired when 2 cell elements in different columns are selected #2');
  checkInputEvent(inputEvents[0], "when 2 cell elements in different columns are selected #2");

  selection.removeAllRanges();
  editor.innerHTML =
    '<table><tr><td id="select" colspan="2">cell1-1</td><td>cell1-3</td></tr><tr><td>cell2-1</td><td>cell2-2</td><td>cell2-3</td></tr></table>';
  beforeInputEvents = [];
  inputEvents = [];
  range = document.createRange();
  range.selectNode(document.getElementById("select"));
  selection.addRange(range);
  getTableEditor().deleteTableColumn(1);
  is(editor.innerHTML, '<table><tbody><tr><td id="select" colspan="1"><br></td><td>cell1-3</td></tr><tr><td>cell2-2</td><td>cell2-3</td></tr></tbody></table>',
     "nsITableEditor.deleteTableColumn(1) with a selected cell is colspan=\"2\" should delete the first column and add empty cell to the second column");
  todo_is(beforeInputEvents.length, 1,
     'Only one "beforeinput" event should be fired when a cell is selected and its colspan is 2');
  if (beforeInputEvents.length > 0) {
    checkInputEvent(beforeInputEvents[0], "when a cell is selected and its colspan is 2");
  }
  is(inputEvents.length, 1,
     'Only one "input" event should be fired when a cell is selected and its colspan is 2');
  checkInputEvent(inputEvents[0], "when a cell is selected and its colspan is 2");

  selection.removeAllRanges();
  editor.innerHTML =
    '<table><tr><td id="select" colspan="3">cell1-1</td></tr><tr><td>cell2-1</td><td>cell2-2</td><td>cell2-3</td></tr></table>';
  beforeInputEvents = [];
  inputEvents = [];
  range = document.createRange();
  range.selectNode(document.getElementById("select"));
  selection.addRange(range);
  getTableEditor().deleteTableColumn(1);
  is(editor.innerHTML, '<table><tbody><tr><td id="select" colspan="2"><br></td></tr><tr><td>cell2-2</td><td>cell2-3</td></tr></tbody></table>',
     "nsITableEditor.deleteTableColumn(1) with a selected cell is colspan=\"3\" should delete the first column and add empty cell whose colspan is 2 to the second column");
  todo_is(beforeInputEvents.length, 1,
     'Only one "beforeinput" event should be fired when a cell is selected and its colspan is 3');
  if (beforeInputEvents.length > 0) {
    checkInputEvent(beforeInputEvents[0], "when a cell is selected and its colspan is 3");
  }
  is(inputEvents.length, 1,
     'Only one "input" event should be fired when a cell is selected and its colspan is 3');
  checkInputEvent(inputEvents[0], "when a cell is selected and its colspan is 3");

  selection.removeAllRanges();
  editor.innerHTML =
    '<table><tr><td colspan="3">cell1-1</td></tr><tr><td>cell2-1</td><td id="select">cell2-2</td><td>cell2-3</td></tr></table>';
  beforeInputEvents = [];
  inputEvents = [];
  range = document.createRange();
  range.selectNode(document.getElementById("select"));
  selection.addRange(range);
  getTableEditor().deleteTableColumn(1);
  is(editor.innerHTML, '<table><tbody><tr><td colspan="2">cell1-1</td></tr><tr><td>cell2-1</td><td>cell2-3</td></tr></tbody></table>',
     "nsITableEditor.deleteTableColumn(1) with selected cell in the second column should delete the second column and the colspan in the first row should be adjusted");
  todo_is(beforeInputEvents.length, 1,
     'Only one "beforeinput" event should be fired when a cell in 2nd column is only cell defined by the column #1');
  if (beforeInputEvents.length > 0) {
    checkInputEvent(beforeInputEvents[0], "when a cell in 2nd column is only cell defined by the column #1");
  }
  is(inputEvents.length, 1,
     'Only one "input" event should be fired when a cell in 2nd column is only cell defined by the column #1');
  checkInputEvent(inputEvents[0], "when a cell in 2nd column is only cell defined by the column #1");

  selection.removeAllRanges();
  editor.innerHTML =
    '<table><tr><td colspan="2">cell1-1</td><td>cell1-3</td></tr><tr><td>cell2-1</td><td id="select">cell2-2</td><td>cell2-3</td></tr></table>';
  beforeInputEvents = [];
  inputEvents = [];
  range = document.createRange();
  range.selectNode(document.getElementById("select"));
  selection.addRange(range);
  getTableEditor().deleteTableColumn(1);
  is(editor.innerHTML, '<table><tbody><tr><td colspan="1">cell1-1</td><td>cell1-3</td></tr><tr><td>cell2-1</td><td>cell2-3</td></tr></tbody></table>',
     "nsITableEditor.deleteTableColumn(1) with selected cell in the second column should delete the second column and the colspan should be adjusted");
  todo_is(beforeInputEvents.length, 1,
     'Only one "beforeinput" event should be fired when a cell in 2nd column is only cell defined by the column #2');
  if (beforeInputEvents.length > 0) {
    checkInputEvent(beforeInputEvents[0], "when a cell in 2nd column is only cell defined by the column #2");
  }
  is(inputEvents.length, 1,
     'Only one "input" event should be fired when a cell in 2nd column is only cell defined by the column #2');
  checkInputEvent(inputEvents[0], "when a cell in 2nd column is only cell defined by the column #2");

  SimpleTest.finish();
});

function getTableEditor() {
  var editingSession = SpecialPowers.wrap(window).docShell.editingSession;
  return editingSession.getEditorForWindow(window).QueryInterface(SpecialPowers.Ci.nsITableEditor);
}

</script>
</body>

</html>
